feat: 기본 팝업과 로그인 팝업 구현
