+++ src/components/Popup/Popup.tsx
@@ -1,0 +1,30 @@
+import Modal from "@/components/Modal";
+import Button from "@/components/Button";
+
+type PopupProps = {
+  isOpen: boolean;
+  type: "alert" | "confirm";
+  onClick: () => void;
+  onClose: () => void;
+  children: React.ReactNode;
+};
+
+export default function Popup({ isOpen, type = "alert", onClose, onClick, children }: PopupProps) {
+  return (
+    <Modal isOpen={isOpen} onClose={onClose} className="w-full max-w-[450px]">
+      <div className="flex h-[151px] w-full flex-col items-center justify-between text-base font-medium">
+        <div className="flex flex-1 flex-col items-center justify-center">{children}</div>
+        <div className="flex gap-2">
+          {type === "confirm" && (
+            <Button styleType="outline" onClick={onClose} className="w-[120px]">
+              취소
+            </Button>
+          )}
+          <Button onClick={onClick} className="w-[120px]">
+            확인
+          </Button>
+        </div>
+      </div>
+    </Modal>
+  );
+}
