+++ src/components/Popup/index.ts
@@ -1,0 +1,2 @@
+export { default as LoginPopup } from "./LoginPopup";
+export { default as Popup } from "./Popup";
